requirements.txt.txt
# Web framework
Flask==3.0.3

# AWS SDK
boto3==1.35.70
botocore==1.35.70

# Data handling
pandas==2.2.3
openpyxl==3.1.5
PyPDF2==3.0.1
python-docx==1.1.2

# PowerPoint handling
python-pptx==0.6.23
pptx2txt2==0.1.4  # If not available, may need to install from source

# Image and PDF processing
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.0.0

# Visualization
graphviz==0.20.3

# Utilities
Werkzeug==3.0.4
shutilwhich==1.1.0  # optional if shutil.which issues occur

# Other possible dependencies (if used in your modules)
requests==2.32.3

# Poppler is required for pdf2image (install via OS package manager: apt-get install poppler-utils on Linux, or download binaries for Windows).